fixed signed-unsigned warnings

+++ src/osgEarth/TextureCompositorTexArray.cpp
@@ -256,10 +256,10 @@ TextureCompositorTexArray::prepareImage( const GeoImage& layerImage, const GeoEx
     if (!image)
         return GeoImage::INVALID;
 
-    if (image->getPixelFormat() != GL_RGBA ||
-        image->getInternalTextureFormat() != GL_RGBA8 ||
-        image->s() != textureSize ||
-        image->t() != textureSize )
+    if (image->getPixelFormat() != (int)GL_RGBA ||
+        image->getInternalTextureFormat() != (int)GL_RGBA8 ||
+        image->s() != (int)textureSize ||
+        image->t() != (int)textureSize )
     {
         // Because all tex2darray layers must be identical in format, let's use RGBA.
         osg::ref_ptr<osg::Image> newImage = ImageUtils::convertToRGBA8( image );
